# REQUEST FOR COMMENTS: Graceful shutdown of active calls