--- a/src/service.js
+++ b/src/service.js
@@ -90,7 +90,9 @@ class Service {
 				// Expose to call `service.actions.find({ ...params })`
 				this.actions[name] = (params, opts) => {
 					const ctx = this.broker.ContextFactory.create(this.broker, innerAction, null, params, opts || {});
-					return innerAction.handler(ctx);
+					return innerAction.handler(ctx).finally(() => {
+						ctx.dispose();
+					});
 				};
 
 			});
@@ -119,22 +121,24 @@ class Service {
 				event.handler = function (payload, sender, eventName) {
 					if (_.isFunction(handler)) {
 						const p = handler.apply(self, [payload, sender, eventName]);
+						// TODO: Track event handler started
 
 						// Handle async-await returns
 						if (utils.isPromise(p)) {
 							/* istanbul ignore next */
 							p.catch(err => self.logger.error(err));
-						}
+						} // TODO: Cleanup event tracking
 
 					} else if (Array.isArray(handler)) {
 						handler.forEach(fn => {
 							const p = fn.apply(self, [payload, sender, eventName]);
+							// TODO: Track event handler started
 
 							// Handle async-await returns
 							if (utils.isPromise(p)) {
 								/* istanbul ignore next */
 								p.catch(err => self.logger.error(err));
-							}
+							} // TODO: Cleanup event tracking
 						});
 					}
 
@@ -192,17 +196,30 @@ class Service {
 		};
 
 		this.stopped = () => {
-			if (_.isFunction(this.schema.stopped))
-				return this.Promise.method(this.schema.stopped).call(this);
+			return new this.Promise((resolve, reject) => {
+				const timeout = setTimeout(reject, 2000);
+				const checkForContexts = () => {
+					if (this._getActiveContexts().size === 0) {
+						clearTimeout(timeout);
+						resolve();
+					} else {
+						setTimeout(checkForContexts, 100);
+					}
+				};
+				setTimeout(checkForContexts, 0);
+			}).finally(() => {
+				if (_.isFunction(this.schema.stopped))
+					return this.Promise.method(this.schema.stopped).call(this);
 
-			if (Array.isArray(this.schema.stopped)) {
-				return this.schema.stopped
-					.reverse()
-					.map(fn => this.Promise.method(fn.bind(this)))
-					.reduce((p, fn) => p.then(fn), this.Promise.resolve());
-			}
+				if (Array.isArray(this.schema.stopped)) {
+					return this.schema.stopped
+						.reverse()
+						.map(fn => this.Promise.method(fn.bind(this)))
+						.reduce((p, fn) => p.then(fn), this.Promise.resolve());
+				}
 
-			return this.Promise.resolve();
+				return this.Promise.resolve();
+			});
 		};
 
 		// Call the created event handler
@@ -258,6 +275,19 @@ class Service {
 		action.metrics = _.defaultsDeep(action.metrics, { params: false, meta: true });
 
 		return action;
+	}
+
+	/**
+	 * Retrieve list of active contexts for the service
+	 * @returns {Set}
+	 * @private
+	 * @memberof Service
+	 */
+	_getActiveContexts() {
+		if (_.isFunction(this.broker.ContextFactory.getActiveContexts)) {
+			return this.broker.ContextFactory.getActiveContexts(this);
+		}
+		return new Set();
 	}
 
 	/**