--- a/test/unit/service-broker.spec.js
+++ b/test/unit/service-broker.spec.js
@@ -115,6 +115,7 @@ describe("Test ServiceBroker constructor", () => {
 			statistics: true,
 			heartbeatTimeout : 20,
 			heartbeatInterval: 5,
+			trackContext: false,
 
 			disableBalancer: true,
 			registry: {