This is an initial step towards a full graceful shutdown of services. Please give comments and ideas for improvements.