@icebob 2018-04-23
This feature is great, but the extra request storage cuts the performance (from 1.9M rps ro 890k rps). 

**Master benchmark:**
```
===============================
  Moleculer common benchmarks
===============================

Platform info:
==============
   Windows_NT 6.1.7601 x64
   Node.JS: 8.11.0
   V8: 6.2.414.50
   Intel(R) Core(TM) i7-4770K CPU @ 3.50GHz × 8

Suite: Local call
√ broker.call (normal)*             1,975,446 rps
√ broker.call (with params)*        2,136,275 rps

   broker.call (normal)* (#)            0%      (1,975,446 rps)   (avg: 506ns)
   broker.call (with params)*       +8.14%      (2,136,275 rps)   (avg: 468ns)
-----------------------------------------------------------------------

Suite: Call with middlewares
√ No middlewares*        1,987,608 rps
√ 5 middlewares*         1,990,242 rps

   No middlewares* (#)       0%      (1,987,608 rps)   (avg: 503ns)
   5 middlewares*        +0.13%      (1,990,242 rps)   (avg: 502ns)
-----------------------------------------------------------------------

Suite: Call with statistics & metrics
√ No statistics*                    1,934,230 rps
√ With metrics*                       578,381 rps
√ With statistics*                    872,608 rps
√ With metrics & statistics*          418,760 rps

   No statistics* (#)                   0%      (1,934,230 rps)   (avg: 517ns)
   With metrics*                    -70.1%        (578,381 rps)   (avg: 1μs)
   With statistics*                -54.89%        (872,608 rps)   (avg: 1μs)
   With metrics & statistics*      -78.35%        (418,760 rps)   (avg: 2μs)
-----------------------------------------------------------------------

Suite: Remote call with FakeTransporter
√ Remote call echo.reply*           49,791 rps

   Remote call echo.reply*           0%         (49,791 rps)   (avg: 20μs)
-----------------------------------------------------------------------
```

**This PR benchmark:**
```
===============================
  Moleculer common benchmarks
===============================

Platform info:
==============
   Windows_NT 6.1.7601 x64
   Node.JS: 8.11.0
   V8: 6.2.414.50
   Intel(R) Core(TM) i7-4770K CPU @ 3.50GHz × 8

Suite: Local call
√ broker.call (normal)*               890,922 rps
√ broker.call (with params)*          912,201 rps

   broker.call (normal)* (#)            0%        (890,922 rps)   (avg: 1μs)
   broker.call (with params)*       +2.39%        (912,201 rps)   (avg: 1μs)
-----------------------------------------------------------------------

Suite: Call with middlewares
√ No middlewares*          886,702 rps
√ 5 middlewares*           883,858 rps

   No middlewares* (#)       0%        (886,702 rps)   (avg: 1μs)
   5 middlewares*        -0.32%        (883,858 rps)   (avg: 1μs)
-----------------------------------------------------------------------

Suite: Call with statistics & metrics
√ No statistics*                      867,774 rps
√ With metrics*                       389,162 rps
√ With statistics*                    514,174 rps
√ With metrics & statistics*          312,388 rps

   No statistics* (#)                   0%        (867,774 rps)   (avg: 1μs)
   With metrics*                   -55.15%        (389,162 rps)   (avg: 2μs)
   With statistics*                -40.75%        (514,174 rps)   (avg: 1μs)
   With metrics & statistics*         -64%        (312,388 rps)   (avg: 3μs)
-----------------------------------------------------------------------

Suite: Remote call with FakeTransporter
√ Remote call echo.reply*           41,863 rps

   Remote call echo.reply*           0%         (41,863 rps)   (avg: 23μs)
-----------------------------------------------------------------------
```