@icebob 2018-04-23
It means if you can't improve the rps above 1.5M, I'll able to merge it only if it's switchable and it's disabled by default.