@rmccallum81 2018-04-26
Ok I'm looking at a few options for increased performance. I'll likely make it switchable anyways but the default will be determined by the performance.