@rmccallum81 2018-04-29
@icebob I've managed to improve the performance as well as add both a switch (default off) and a performance test