@zllovesuki 2018-04-29
Not sure if measuring performance purely based on rps on a local machine is a good metrics. After all, they are microserives not necessary on a single machine. Maybe benchmark it with a networked transporter and see if the performance differs a lot (yes trackContext vs no trackContext)?