@rmccallum81 2018-04-30
The benchmark test is to determine the difference in performance based on constructing, storing, recalling, and disposing the context. By doing it locally only you eliminate variances like network latency which would lead to inconsistent results.