@rmccallum81 2018-04-30
I looked into the thought @zllovesuki expressed about local vs remote performance with the context tracking. I found that when the FakeTransporter was used there was only a ~6% decrease in performance. Not sure if this is a representation of a need to look into the test regarding the transporter or something else since that benchmark is already significantly slower than the pure local calls. I've added a benchmark test surrounding this scenario.